// src/app/enums.ts
export enum Endpoint {
    login = '/auth/login',
    signup = '/auth/register',
    carBrands = '/CarBrands',
    government = '/helper/governorates',
    brands = '/carBrands',
    carsModels = '/CarModels',
    motors = '/motorcycleBrands',
    geerBox = '/cars/gearBoxes',
    shapeTypes = '/cars/shapeTypes',
    adsTypes = '/ads/types',
    carModels = '/CarBrands/:id',
    newCarAds = '/CarAds'
}
export enum baseURL {
    baseURL = 'https://api20200403105050.azurewebsites.net/api'
} 